DAL layer refactored
