+++ Training.Core/Models/RoleClaim.cs
@@ -1,17 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 
-#nullable disable
-
-namespace Training.DAL.Auth
+namespace Training.Core.Models
 {
-    public partial class UserClaim
+    public partial class RoleClaim
     {
         public int Id { get; set; }
-        public Guid UserId { get; set; }
+        public Guid RoleId { get; set; }
         public string ClaimType { get; set; }
         public string ClaimValue { get; set; }
 
-        public virtual User User { get; set; }
+        public virtual Role Role { get; set; }
     }
 }
