+++ Training.Core/Models/UserClaim.cs
@@ -1,17 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 
-#nullable disable
-
-namespace Training.DAL.Auth
+namespace Training.Core.Models
 {
-    public partial class RoleClaim
+    public partial class UserClaim
     {
         public int Id { get; set; }
-        public Guid RoleId { get; set; }
+        public Guid UserId { get; set; }
         public string ClaimType { get; set; }
         public string ClaimValue { get; set; }
 
-        public virtual Role Role { get; set; }
+        public virtual UserAuth User { get; set; }
     }
 }
