+++ Training.DAL/Implementations/UserRepository.cs
@@ -1,18 +1,17 @@
 ﻿using System.Linq;
 using Training.Core.Models;
 using Training.Core.Repositories;
+using Training.DAL.Base;
+using Training.DAL.Context;
 
-namespace Training.DAL
+namespace Training.DAL.Implementations
 {
-    public class UserRepository : IUserRepository
+    public class UserRepository : RepositoryBase, IUserRepository
     {
 
-        private readonly TrainingDbContext _trainingDbContext;
-
-        public UserRepository(TrainingDbContext context)
-        {
-            _trainingDbContext = context;
-        }
+        public UserRepository(TrainingDbContext trainingDbContext)
+           : base(trainingDbContext)
+        { }
 
         public User Get(string username, string password)
         {
